fix(lf) : fix to use z with white space name dir
With `lf -remote`, cd executes in lf and /d option doesn't work like
cmd.
instead of /d, using absolute path can go to other drive with just cd.

but if the directory has white space, lf cannot recognize the directory
so I add !z_dir! with '' to deal with directory which has white space.

But the remaining problem is that when I use 'z' in directory which has
white space in name, it doesn't work.