fix(lf) : fix zoxide error from directory with white space
The problem is that we cannot goto other directory from location with
white space.
it is because `--exclude <path>` option of `zoxide query` has error when
the <path> has white space. so zoxide cannot find the keyword in query.

<path> must be enclosed by "". because in cmd prompt, directory with
white space deal with double quotes not single quotes.

All done:)