feat(komorebi) : add some managed/float rules
float_rules means ignore_rules. Programs which is registered in
float_rules don't be managed by komorebi.
So they persists across multiple workspace.
Of course, it would be nice if the floating windows could be managed by each workspace
individually, but if that is not possible, that is nothing
because managed window has priority over than them.