fix(vieb) : replace vimcommand to nvim in new floating cmd window
if you set `nvim` as vimcommand, the instance is opened in current cmd
prompt. It makes nvim invisible although other gui app such as nvim-qt
or gvim are visible.

so you need to add `start` to open nvim in new window.
the temp filename is *.txt, so you need to change filetype to markdown
for convenient writing usage.