fix(download_and_run) : support command option to cmd_map
1) Problem

If sioyek pref configuration is set to reuse window, the file open in
the same instance which is opened already.
It can makes conflicts. -> add `--new-instance` option

2) Solution

subprocess.run() accepts list form of command. not string that includes
white space. so it needs to change string form of command to list.