fix(clink-prompt) : bug fix and improve git status display
1) Problem

If you add some files to commit, it shows with 'M1' not 'A1'
because we use 'A1' when status message has  '1 A.' that means new file
is added, not tracked file.

It could be confusing to what is added. because I want to consider
both new file and tracked file as added mark.

Also, When I edit some files in working directory, the possible state is
only 'modified' and 'deleted', not 'renamed', 'added', 'copied'.
So These three marks are discarded.

2) Solution

- Removed 'renamed' , 'added', 'copied' checking for working directory
change.

- Distinguish 'renamed', 'copied' from 'added'. and merge 'added',
  'modified' to 'A' mark. renamed and copied are shown after commit.

- Add 'U' mark for merge conflicted after commit.